refactor: move files around to for better organization
