+++ apps/backend/src/application/dao/geolocation.service.abstract.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable } from '@nestjs/common';
 
 interface ReversedGeocodingOutput {
   region: string;
