--- apps/backend/src/auth/auth.module.ts
@@ -1,16 +1,0 @@
-import { Module } from '@nestjs/common';
-import { JwtModule } from '@nestjs/jwt';
-
-import { UserModule } from '../user/user.module';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
-
-import { AccessTokenStrategy } from './strategies/accessToken.strategy';
-import { RefreshTokenStrategy } from './strategies/refreshToken.strategy';
-
-@Module({
-  imports: [JwtModule.register({}), UserModule],
-  controllers: [AuthController],
-  providers: [AuthService, AccessTokenStrategy, RefreshTokenStrategy],
-})
-export class AuthModule {}
