+++ apps/backend/src/modules/auth.module.ts
@@ -1,0 +1,14 @@
+import { Module } from '@nestjs/common';
+import { JwtModule } from '@nestjs/jwt';
+import { AuthService } from '../auth/auth.service';
+import { AccessTokenStrategy } from '../auth/strategies/accessToken.strategy';
+import { RefreshTokenStrategy } from '../auth/strategies/refreshToken.strategy';
+import { UserModule } from './user.module';
+import { AuthController } from '../infra/controllers/auth.controller';
+
+@Module({
+  imports: [JwtModule.register({}), UserModule],
+  controllers: [AuthController],
+  providers: [AuthService, AccessTokenStrategy, RefreshTokenStrategy],
+})
+export class AuthModule {}
