+++ apps/backend/src/modules/user.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UserService } from './user.service';
+import { UserService } from '../user/user.service';
 
 @Module({
   providers: [UserService],
